server/python_backend/requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
python-dotenv==1.0.0
pypdf==3.17.1
pdfplumber==0.10.3 